+++ main.hpp
@@ -52,7 +52,6 @@ extern std::map<std::string, serial_val> set_attributes;
 extern long long extend_zooms_max;
 extern int retain_points_multiplier;
 extern size_t maximum_string_attribute_length;
-extern std::string accumulate_numeric;
 extern unsigned long long preserve_multiplier_density_threshold;
 
 struct order_field {