+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 2.80.0
+
+* Remove undocumented command-line options
+
+# 2.79.0
+
+* When deduplicating features by ID in tippecanoe-overzoom, be careful
+  to track even features that have been clipped away.
+
 # 2.78.0
 
 * Fix potential infinite loops in as-needed dropping and coalescing.